+++ Example/Emission/ARRootViewController.m
@@ -135,7 +135,7 @@
 - (ARCellData *)jumpToArtist
 {
   return [self tappableCellDataWithTitle:@"Artist" selection: ^{
-    id viewController = [[ARArtistComponentViewController alloc] initWithArtistID:@"david-shrigley"];
+    id viewController = [[ARArtistComponentViewController alloc] initWithArtistID:@"marina-abramovic-1"];
     [self.navigationController pushViewController:viewController animated:YES];
   }];
 }
@@ -203,7 +203,10 @@
   return [self tappableCellDataWithTitle:@"Start Consignment Flow" selection:^{
     [[(EigenLikeNavigationController *)self.navigationController backButton] setHidden:YES];
     id viewController = [[ARComponentViewController alloc] initWithEmission: nil moduleName:@"Consignments" initialProperties: @{}];
-    [self.navigationController pushViewController:viewController animated:YES];
+
+    UINavigationController *nav = [[UINavigationController alloc] initWithRootViewController:viewController];
+    nav.navigationBarHidden = YES;
+    [self.navigationController presentViewController:nav animated:YES completion:NULL];
   }];
 }
 
@@ -245,6 +248,28 @@
   return crashCellData;
 }
 
+- (ARCellData *)toggleRNPSwitch
+{
+  BOOL forceRNP = [[NSUserDefaults standardUserDefaults] boolForKey:ARForceUseRNPDefault];
+  NSString *rnpLocation = [[NSUserDefaults standardUserDefaults] stringForKey:ARRNPackagerHostDefault];
+  NSString *title = !forceRNP ? [NSString stringWithFormat:@"Use RNP at %@", rnpLocation] : @"Revert forced RNP";
+
+  ARCellData *crashCellData = [[ARCellData alloc] initWithIdentifier:AROptionCell];
+  [crashCellData setCellConfigurationBlock:^(UITableViewCell *cell) {
+    cell.textLabel.text = title;
+  }];
+
+  [crashCellData setCellSelectionBlock:^(UITableView *tableView, NSIndexPath *indexPath) {
+    [self showAlertViewWithTitle:@"Confirm Switch" message:@"Switching forced RNP settings." actionTitle:@"Continue" actionHandler:^{
+
+      [[NSUserDefaults standardUserDefaults] setBool:!forceRNP forKey:ARForceUseRNPDefault];
+      [[NSUserDefaults standardUserDefaults] synchronize];
+      exit(0);
+    }];
+  }];
+  return crashCellData;
+}
+
 - (ARCellData *)emissionJSLocationDescription:(NSString *)loadedFromString
 {
   ARCellData *crashCellData = [[ARCellData alloc] initWithIdentifier:AROptionCell];
@@ -271,14 +296,18 @@
 {
   ARSectionData *sectionData = [[ARSectionData alloc] init];
   [self setupSection:sectionData withTitle:@"Admin"];
-
-  if ([AppSetup ambientSetup].inStaging) {
+  AppSetup *setup = [AppSetup ambientSetup];
+  if (setup.inStaging) {
     [sectionData addCellDataFromArray:@[
       [self editableTextCellDataWithName:@"Gravity API" defaultKey:ARStagingAPIURLDefault],
       [self editableTextCellDataWithName:@"Metaphysics API" defaultKey:ARStagingMetaphysicsURLDefault],
       [self editableTextCellDataWithName:@"RN Packager" defaultKey:ARRNPackagerHostDefault],
     ]];
+    if (!setup.inSimulator) {
+      [sectionData addCellData:self.toggleRNPSwitch];
+    }
   }
+
 
   [sectionData addCellData:self.generateStagingSwitch];
   [sectionData addCellData:self.logOutButton];
@@ -295,6 +324,8 @@
     }];
   }];
 }
+
+
 
 
 @end