+++ src/lib/Components/Consignments/Screens/Metadata.tsx
@@ -1,6 +1,16 @@
 import * as React from "react"
 
-import { LayoutAnimation, NavigatorIOS, Picker, Route, ScrollView, View, ViewProperties } from "react-native"
+import {
+  LayoutAnimation,
+  NavigatorIOS,
+  Picker,
+  Route,
+  ScrollView,
+  TouchableHighlight,
+  TouchableWithoutFeedback,
+  View,
+  ViewProperties,
+} from "react-native"
 import ConsignmentBG from "../Components/ConsignmentBG"
 import { BodyText, LargeHeadline, Subtitle } from "../Typography"
 
@@ -78,8 +88,8 @@ export default class Metadata extends React.Component<Props, State> {
     return (
       <ConsignmentBG>
         <DoneButton onPress={this.doneTapped}>
-          <ScrollView>
-            <View>
+          <ScrollView keyboardShouldPersistTaps="always">
+            <View style={{ padding: 10 }}>
               <Row>
                 <Text
                   text={{ placeholder: "Title", onChangeText: this.updateTitle, value: this.state.title }}
@@ -95,15 +105,17 @@ export default class Metadata extends React.Component<Props, State> {
               </Row>
 
               <Row>
-                <Text
-                  text={{
-                    placeholder: "Category",
-                    onChangeText: this.updateCategory,
-                    onFocus: this.showCategorySelection,
-                    value: this.state.category,
-                  }}
-                  style={{ margin: 10 }}
-                />
+                <TouchableWithoutFeedback onPress={() => console.log("HI")} style={{ flex: 1, height: 40 }}>
+                  <Text
+                    text={{
+                      placeholder: "Category",
+                      onChangeText: this.updateCategory,
+                      editable: false,
+                      value: this.state.category,
+                    }}
+                    style={{ margin: 10 }}
+                  />
+                </TouchableWithoutFeedback>
               </Row>
 
               {this.state.showSelector &&
@@ -139,10 +151,18 @@ export default class Metadata extends React.Component<Props, State> {
               </Row>
 
               <Row>
-                <Text text={{ placeholder: "Depth", onChange: this.updateDepth }} style={{ margin: 10 }} />
-                <View style={{ flex: 1, flexDirection: "row", alignContent: "center", margin: 10 }}>
+                <Text
+                  text={{
+                    keyboardType: "numeric",
+                    placeholder: "Depth",
+                    onChangeText: this.updateDepth,
+                    value: this.state.depth ? this.state.depth.toString() : "",
+                  }}
+                  style={{ margin: 10 }}
+                />
+                <View style={{ flex: 1, flexDirection: "row", alignItems: "center", margin: 10 }}>
                   <Label>Units</Label>
-                  <Toggle selected={this.state.unit === "cm"} left="CM" right="IN" onPress={this.updateUnit} />
+                  <Toggle selected={this.state.unit === "CM"} left="CM" right="IN" onPress={this.updateUnit} />
                 </View>
               </Row>
             </View>