+++ src/lib/Components/Consignments/Submission/objectToGraphQL.ts
@@ -14,6 +14,9 @@ export const objectToGraphQLInput = (obj: any, enums?: string[]) => {
         input += key + `: ` + element + `, `
       } else if (isString(element)) {
         input += key + `: "` + element + `", `
+      } else if (element) {
+        // is probably an object at this point
+        input += key + `: ` + objectToGraphQLInput(element) + ", "
       }
     }
   }