+++ src/lib/Components/Home/ArtworkRails/ArtworkRailHeader.tsx
@@ -17,6 +17,7 @@ const { ARTemporaryAPIModule } = NativeModules
 import Events from "../../../NativeModules/Events"
 
 import colors from "../../../../data/colors"
+import fonts from "../../../../data/fonts"
 import Button from "../../Buttons/InvertedButton"
 import SerifText from "../../Text/Serif"
 import SectionTitle from "../SectionTitle"
@@ -133,11 +134,10 @@ const styles = StyleSheet.create<Styles>({
   title: {
     marginTop: 10,
     fontSize: isPad ? 30 : 26,
-    alignSelf: "center",
-    textAlign: "center",
+    textAlign: "left",
   },
   viewAllButton: {
-    fontFamily: "Avant Garde Gothic ITCW01Dm",
+    fontFamily: fonts["avant-garde-regular"],
     fontSize: isPad ? 14 : 12,
     color: colors["gray-medium"],
     textAlign: "center",